+++ lib/logstash/inputs/dead_letter_queue.rb
@@ -66,7 +66,14 @@ class LogStash::Inputs::DeadLetterQueue < LogStash::Inputs::Base
       # clean_consumed requires the commit of offset
       raise LogStash::ConfigurationError.new("enabling clean_consumed requires commit_offsets to also be enabled")
     end
-    @inner_plugin = org.logstash.input.DeadLetterQueueInputPlugin.new(dlq_path, @commit_offsets, sincedb_path, start_timestamp, clean_consumed)
+    @cleaned_metrics = metric.namespace(@pipeline_id)
+    @inner_plugin = org.logstash.input.DeadLetterQueueInputPlugin.new(dlq_path, @commit_offsets, sincedb_path, start_timestamp, clean_consumed,
+            lambda do |segments, events|
+                # gauges is used instead of metric type because the updates that comes from the
+                # DLQ reader are already absolute values and not deltas.
+                @cleaned_metrics.gauge(:cleaned_segments, segments)
+                @cleaned_metrics.gauge(:cleaned_events, events)
+            end)
     @inner_plugin.register
 
     if Gem::Requirement.new('< 7.0').satisfied_by?(logstash_version)